bin/death_sentence_ble-main 2/requirements.txt
# Core dependencies for the DeathScent UI
flask>=2.0.0
flask-cors>=3.0.0
bleak>=0.19.0

# Optional dependencies (for legacy code compatibility)
crcmod>=1.7
pyserial>=3.5

# macOS-specific dependencies (if on macOS)
pyobjc-core>=11.1
pyobjc-framework-Cocoa>=11.1
pyobjc-framework-CoreBluetooth>=11.1
pyobjc-framework-libdispatch>=11.1

# Legacy SDK (no longer used in new implementation)
# ScentRealmForNeckWear==1.0.2